README.md updated for .env file explanation + python bot logging & split
The bot wasn't working as expected. It was du to the way the .env file was formatted.

I've added some logging in the python script and I've edited the README.md to reflect the needed changes in the .env file for the bot to work correctly.

I also had to adapt the split() to split(',') in the python script to match with the way the .env file is formatted.